スナックバーによるアップロード候補表示をコンポーネントに分割
